Create shared CtaCard component
