+++ src/modules/explore/pages/exploreDaosPage/exploreDaosPageClient.tsx
@@ -10,10 +10,7 @@ import { Carousel } from '@/shared/components/carousel';
 import { Container } from '@/shared/components/container';
 import { useDialogContext } from '@/shared/components/dialogProvider';
 import { useTranslations } from '@/shared/components/translationsProvider';
-import doItYourselfIcon from '../../../../assets/images/doItYourselfIcon.svg';
-import enterpriseServiceIcon from '../../../../assets/images/enterpriseServiceIcon.svg';
 import NetBackground from '../../../../assets/images/net_bg.svg';
-import noCodeSetupIcon from '../../../../assets/images/noCodeSetup.svg';
 import type { IGetDaoListParams } from '../../api/daoExplorerService';
 import { CtaCard } from '../../components/ctaCard';
 import { DaoCarouselCard } from '../../components/daoCarouselCard';
@@ -96,18 +93,22 @@ export const ExploreDaosPageClient: React.FC<IExploreDaosPageClientProps> = (
                     <ExploreSection
                         title={t('app.explore.exploreDaosPage.section.cta')}
                     >
-                        <div className="flex flex-col items-start gap-4 self-stretch md:flex-row md:gap-4 lg:gap-8">
+                        <div className="flex flex-col gap-4 self-stretch md:flex-row md:gap-4 lg:gap-8">
                             <CtaCard
                                 actionHref="https://www.aragon.org/get-assistance-form"
                                 actionLabel={t(
                                     'app.explore.exploreDaosPage.enterpriseService.actionLabel',
                                 )}
-                                imgSrc={enterpriseServiceIcon as string}
-                                isExternal={true}
-                                isPrimary={true}
-                                subtitle={t(
+                                className="flex-1"
+                                description={t(
                                     'app.explore.exploreDaosPage.enterpriseService.subtitle',
                                 )}
+                                isPrimary={true}
+                                objectType="SMART_CONTRACT"
+                                tag={t(
+                                    'app.explore.exploreDaosPage.enterpriseService.tag',
+                                )}
+                                textSize="smaller"
                                 title={t(
                                     'app.explore.exploreDaosPage.enterpriseService.title',
                                 )}
@@ -119,12 +120,13 @@ export const ExploreDaosPageClient: React.FC<IExploreDaosPageClientProps> = (
                                 actionOnClick={() =>
                                     open(CreateDaoDialogId.CREATE_DAO_DETAILS)
                                 }
-                                imgSrc={noCodeSetupIcon as string}
-                                isExternal={false}
-                                isPrimary={false}
-                                subtitle={t(
+                                className="flex-1"
+                                description={t(
                                     'app.explore.exploreDaosPage.noCodeSetup.subtitle',
                                 )}
+                                isPrimary={false}
+                                objectType="APP"
+                                textSize="smaller"
                                 title={t(
                                     'app.explore.exploreDaosPage.noCodeSetup.title',
                                 )}
@@ -134,12 +136,13 @@ export const ExploreDaosPageClient: React.FC<IExploreDaosPageClientProps> = (
                                 actionLabel={t(
                                     'app.explore.exploreDaosPage.doItYourself.actionLabel',
                                 )}
-                                imgSrc={doItYourselfIcon as string}
-                                isExternal={true}
-                                isPrimary={false}
-                                subtitle={t(
+                                className="flex-1"
+                                description={t(
                                     'app.explore.exploreDaosPage.doItYourself.subtitle',
                                 )}
+                                isPrimary={false}
+                                objectType="BOOK"
+                                textSize="smaller"
                                 title={t(
                                     'app.explore.exploreDaosPage.doItYourself.title',
                                 )}
