+++ src/shared/components/ctaCard/index.ts
@@ -1,0 +1,1 @@
+export { CtaCard, type ICtaCardProps } from './ctaCard';
